// src/lib/types.ts
export interface Organization {
  id: string;
  name: string;
  description: string;
  members: string[];
  groups: string[];
  createdBy: string;
  createdAt: string;
}

export interface OrganizationInput {
  name: string;
  description: string;
  members: string[];
}

export interface OrganizationUpdateInput {
  name?: string;
  description?: string;
  members?: string[];
}